% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{available_algorithms}
\alias{available_algorithms}
\title{Lists all algorithms currently supported}
\usage{
available_algorithms()
}
\value{
a list of all available parsnip engines
}
\description{
Lists all algorithms currently supported
}
\examples{
available_algorithms()
}
