% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{available_evaluation_metrics}
\alias{available_evaluation_metrics}
\title{Lists all evaluation metrics currently supported}
\usage{
available_evaluation_metrics()
}
\value{
a list of all available evaluation metrics and their implementation in functional form
}
\description{
Lists all evaluation metrics currently supported
}
\examples{
available_evaluation_metrics()
}
