% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_f1_scores}
\alias{calculate_f1_scores}
\title{Calculate F1 scores}
\usage{
calculate_f1_scores(y, yhat)
}
\arguments{
\item{y}{character or boolean vector of target values}

\item{yhat}{character or boolean vector of predicted values}
}
\value{
named vector of floats
}
\description{
Calculate F1 scores
}
