% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{.eval}
\alias{.eval}
\title{Generic evaluation function skeleton}
\usage{
.eval(y, yhat)
}
\arguments{
\item{y}{vector of target values}

\item{yhat}{vector of predicted values}
}
\description{
Generic evaluation function skeleton
}
