% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{eval_binary_f1_score}
\alias{eval_binary_f1_score}
\title{Evaluate binary F1}
\usage{
eval_binary_f1_score(y, yhat, positive.class)
}
\arguments{
\item{y}{character or boolean vector of target values}

\item{yhat}{character or boolean vector of predicted values}

\item{positive.class}{character scalar for the name of the positive class}
}
\value{
float scalar representing the f1 score
}
\description{
Evaluate binary F1
}
\examples{
y = rbinom(100, 1, prob = 0.2)
yhat = sample(y, replace = TRUE)
eval_binary_f1_score(y, yhat, positive.class = '1')

y = as.logical(rbinom(100, 1, prob = 0.2))
yhat = sample(y, replace = TRUE)
eval_binary_f1_score(y, yhat, positive.class = '1')
}
