% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{eval_f1_score}
\alias{eval_f1_score}
\title{Evaluate F1
Provides a wrapper around underlying functions that calculate either the
binary or the multiclass F1 scores, depending on the target data.}
\usage{
eval_f1_score(y, yhat, positive.class = "1", average = "weighted")
}
\arguments{
\item{y}{character or boolean vector of target values}

\item{yhat}{character or boolean vector of predicted values}

\item{positive.class}{character scalar for the name of the positive class}

\item{average}{character scalar denoting the approach to average multiclass f1 scores}
}
\value{
float scalar representing the f1 score
}
\description{
Evaluate F1
Provides a wrapper around underlying functions that calculate either the
binary or the multiclass F1 scores, depending on the target data.
}
\examples{
y = letters
yhat = sample(y, replace = TRUE)
eval_f1_score(y, yhat)

y = rbinom(100, 1, prob = 0.2)
yhat = sample(y, replace = TRUE)
eval_f1_score(y, yhat)
}
