% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{eval_macro_f1_score}
\alias{eval_macro_f1_score}
\title{Evaluate macro F1 score (unweighted)}
\usage{
eval_macro_f1_score(y, yhat)
}
\arguments{
\item{y}{character or boolean vector of target values}

\item{yhat}{character or boolean vector of predicted values}
}
\value{
float scalar representing the f1 score
}
\description{
Evaluate macro F1 score (unweighted)
}
\examples{
y = letters
yhat = sample(y, replace = TRUE)
eval_f1_score(y, yhat)
}
