% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{eval_mae}
\alias{eval_mae}
\title{Evaluate Mean Absolute Error}
\usage{
eval_mae(y, yhat)
}
\arguments{
\item{y}{vector of target values}

\item{yhat}{vector of predicted values}
}
\value{
float scalar
}
\description{
Evaluate Mean Absolute Error
}
\examples{
y = c(1:10)
yhat = y + rnorm(n = length(y))
eval_mae(y, yhat)
}
