% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{eval_rmse}
\alias{eval_rmse}
\title{Evaluate Root Mean Squared Error}
\usage{
eval_rmse(y, yhat)
}
\arguments{
\item{y}{vector of target values}

\item{yhat}{vector of predicted values}
}
\value{
float scalar
}
\description{
Evaluate Root Mean Squared Error
}
\examples{
y = c(1:10)
yhat = y + rnorm(n = length(y))
eval_rmse(y, yhat)
}
