% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{eval_weighted_f1_score}
\alias{eval_weighted_f1_score}
\title{Evaluate weighted F1 score}
\usage{
eval_weighted_f1_score(y, yhat)
}
\arguments{
\item{y}{character or boolean vector of target values}

\item{yhat}{character or boolean vector of predicted values}
}
\value{
float scalar representing the f1 score
}
\description{
Evaluate weighted F1 score
}
\examples{
y = letters
yhat = sample(y, replace = TRUE)
eval_f1_score(y, yhat)
}
