% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{evaluation_functions}
\alias{evaluation_functions}
\title{Lists all evaluation functions currently supported}
\usage{
evaluation_functions()
}
\value{
list of evaluation functions
}
\description{
Lists all evaluation functions currently supported
}
\examples{
evaluation_functions()
}
