% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{normalize_score}
\alias{normalize_score}
\title{Normalizes the original score compared to a naive baseline score
The calculation that's being performed depends on the type of model}
\usage{
normalize_score(score_baseline, score_original, type)
}
\arguments{
\item{score_baseline}{float, the evaluation metric score for a naive model}

\item{score_original}{float, the evaluation metric score for a statistical model}

\item{type}{character, type of model}
}
\value{
float, normalized score
}
\description{
Normalizes the original score compared to a naive baseline score
The calculation that's being performed depends on the type of model
}
