% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppsr.R
\docType{package}
\name{ppsr}
\alias{ppsr}
\title{ppsr: An R implementation of the Predictive Power Score (PPS)}
\description{
The PPS is an asymmetric, data-type-agnostic score that can detect linear or
non-linear relationships between two columns. The score ranges from 0
(no predictive power) to 1 (perfect predictive power). It can be used as an
alternative to the correlation (matrix).
}
