% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Calculate predictive power score for x on y}
\usage{
score(
  df,
  x,
  y,
  algorithm = "tree",
  metrics = list(regression = "MAE", classification = "F1_weighted"),
  cv_folds = 5,
  seed = 1
)
}
\arguments{
\item{df}{data.frame containing columns for x and y}

\item{x}{character, column name of predictor variable}

\item{y}{character, column name of target variable}

\item{algorithm}{character}

\item{metrics}{named list of \code{eval_*} functions used for
regression and classification problems}

\item{cv_folds}{float, number of cross-validation folds}

\item{seed}{float, seed to ensure reproducibility/stability}
}
\value{
float,  representing predictive power score
}
\description{
Calculate predictive power score for x on y
}
\examples{
score(iris, x = 'Petal.Length', y = 'Species')
}
