% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_correlations}
\alias{score_correlations}
\title{Calculate correlation coefficients for whole dataframe}
\usage{
score_correlations(df, ...)
}
\arguments{
\item{df}{data.frame containing columns for x and y}

\item{...}{arguments to pass to \code{stats::cor()}}
}
\value{
data.frame with x-y correlation coefficients
}
\description{
Calculate correlation coefficients for whole dataframe
}
\examples{
score_correlations(iris)
}
