% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_df}
\alias{score_df}
\title{Calculate predictive power scores for whole dataframe
Iterates through the columns of the dataframe, calculating the predictive power
score for every possible combination of \code{x} and \code{y}.}
\usage{
score_df(df, ...)
}
\arguments{
\item{df}{data.frame containing columns for x and y}

\item{...}{any arguments passed to \code{\link{score}}}
}
\value{
dataframe, detailed report on predictive power scores
}
\description{
Calculate predictive power scores for whole dataframe
Iterates through the columns of the dataframe, calculating the predictive power
score for every possible combination of \code{x} and \code{y}.
}
\examples{
score_df(iris)
}
