% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_matrix}
\alias{score_matrix}
\title{Calculate predictive power score matrix
Iterates through the columns of the dataset, calculating the predictive power
score for every possible combination of \code{x} and \code{y}.}
\usage{
score_matrix(df, ...)
}
\arguments{
\item{df}{data.frame containing columns for x and y}

\item{...}{any arguments passed to \code{\link{score_df}},
some of which will be passed on to \code{\link{score}}}
}
\value{
a matrix of numeric values, representing predictive power scores
}
\description{
Note that the targets are on the rows, and the features on the columns.
}
\examples{
\donttest{score_matrix(iris)}
\donttest{score_matrix(mtcars, do_parallel = TRUE, n_cores=2)}
}
