% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_model}
\alias{score_model}
\title{Calculates out-of-sample model performance of a statistical model}
\usage{
score_model(train, test, model, eval_fun)
}
\arguments{
\item{train}{df, training data, containing variable y}

\item{test}{df, test data, containing variable y}

\item{model}{parsnip model object, with mode preset}

\item{eval_fun}{function, evaluation function being used}
}
\value{
float, evaluation score for predictions using naive model
}
\description{
Calculates out-of-sample model performance of a statistical model
}
