% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Calculate predictive power score for x on y}
\usage{
score(x = mtcars$mpg, y = mtcars$cyl, algorithm = "tree", metric = mae)
}
\arguments{
\item{x}{any type of vector of predictor values}

\item{y}{any type of vector of target values}

\item{algorithm}{character scalar}

\item{metric}{model evaluation function that takes arguments \code{y} and \code{yhat} and returns a numeric scalar}
}
\value{
numeric scalar representing predictive power score
}
\description{
Calculate predictive power score for x on y
}
\examples{
score(x = mtcars$mpg, y = mtcars$cyl, algorithm = 'tree', metric = mae)
}
