% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Calculate predictive power score for x on y}
\usage{
score(
  x,
  y,
  algorithm = "tree",
  eval_funs = list(regression = eval_mae, classification = eval_f1_score)
)
}
\arguments{
\item{x}{any type of vector of predictor values}

\item{y}{any type of vector of target values}

\item{algorithm}{character scalar}

\item{eval_funs}{named list of \code{eval_*} functions used for
regression and classification problems}
}
\value{
numeric scalar representing predictive power score
}
\description{
Calculate predictive power score for x on y
}
\examples{
score(x = mtcars$mpg, y = mtcars$cyl)

score(x = iris$Petal.Length, y = iris$Species)
}
