% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_df}
\alias{score_df}
\title{Calculate predictive power scores for whole dataframe
Iterates through the columns of the dataframe, calculating the predictive power
score for every possible combination of \code{x} and \code{y}.}
\usage{
score_df(df, ..., do_parallel = FALSE, nc = -1)
}
\arguments{
\item{df}{data.frame containing columns for x and y}

\item{...}{any arguments passed to \code{\link{score}}}

\item{do_parallel}{bool, whether to perform \code{\link{score}} calls in parallel}

\item{nc}{numeric, number of cores to use, defaults to maximum minus 1}
}
\value{
dataframe, detailed report on predictive power scores
}
\description{
Calculate predictive power scores for whole dataframe
Iterates through the columns of the dataframe, calculating the predictive power
score for every possible combination of \code{x} and \code{y}.
}
\examples{
score_df(iris)
\dontrun{score_df(mtcars, do_parallel = TRUE)}
}
