% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_model}
\alias{score_model}
\title{Calculates out-of-sample model performance of a statistical model}
\usage{
score_model(train, test, model, x, y, metric)
}
\arguments{
\item{train}{df, training data, containing variable y}

\item{test}{df, test data, containing variable y}

\item{model}{parsnip model object, with mode preset}

\item{x}{character, column name of predictor variable}

\item{y}{character, column name of target variable}

\item{metric}{character, name of evaluation metric being used, see \code{available_evaluation_metrics()}}
}
\value{
numeric vector of length one, evaluation score for predictions using naive model
}
\description{
Calculates out-of-sample model performance of a statistical model
}
