% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_naive}
\alias{score_naive}
\title{Calculate out-of-sample model performance of naive baseline model
The calculation that's being performed depends on the type of model
For regression models, the mean is used as prediction
For classification, a model predicting random values and
a model predicting modal values are used and
the best model is taken as baseline score}
\usage{
score_naive(train, test, x, y, type, metric)
}
\arguments{
\item{train}{df, training data, containing variable y}

\item{test}{df, test data, containing variable y}

\item{x}{character, column name of predictor variable}

\item{y}{character, column name of target variable}

\item{type}{character, type of model}

\item{metric}{character, evaluation metric being used}
}
\value{
float, evaluation score for predictions using naive model
}
\description{
Calculate out-of-sample model performance of naive baseline model
The calculation that's being performed depends on the type of model
For regression models, the mean is used as prediction
For classification, a model predicting random values and
a model predicting modal values are used and
the best model is taken as baseline score
}
