% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_predictors}
\alias{score_predictors}
\title{Calculate predictive power scores for y}
\usage{
score_predictors(df, y, ...)
}
\arguments{
\item{df}{data.frame}

\item{y}{character scalar with target column name}

\item{...}{any arguments passed to \code{\link{score}}}
}
\value{
list of float scalars, representing predictive power scores
}
\description{
Calculate predictive power scores for y
}
\examples{
score_predictors(df = mtcars, y = 'mpg')
}
