% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_predictors}
\alias{score_predictors}
\title{Calculate predictive power scores for y
Calculates the predictive power scores for the specified \code{y} variable
using every column in the dataset as \code{x}, including itself.}
\usage{
score_predictors(df, y, ..., do_parallel = FALSE, n_cores = -1)
}
\arguments{
\item{df}{data.frame containing columns for x and y}

\item{y}{character, column name of target variable}

\item{...}{any arguments passed to \code{\link{score}}}

\item{do_parallel}{bool, whether to perform \code{\link{score}} calls in parallel}

\item{n_cores}{numeric, number of cores to use, defaults to maximum minus 1}
}
\value{
dataframe, detailed report on predictive power scores
}
\description{
Calculate predictive power scores for y
Calculates the predictive power scores for the specified \code{y} variable
using every column in the dataset as \code{x}, including itself.
}
\examples{
\dontrun{score_predictors(df = iris, y = 'Species')}
\dontrun{score_predictors(df = mtcars, y = 'mpg', do_parallel = TRUE)}
}
