% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_matrix}
\alias{visualize_matrix}
\title{Visualize the PPS matrix}
\usage{
visualize_matrix(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
ggplot2 heatmap visualization
}
\description{
Visualize the PPS matrix
}
\examples{
visualize_matrix(mtcars)
}
