% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_matrix}
\alias{visualize_matrix}
\title{Visualize the PPS matrix}
\usage{
visualize_matrix(df, color = "#08306B", ...)
}
\arguments{
\item{df}{data.frame containing columns for x and y}

\item{color}{color used for highlighting high PPS}

\item{...}{any arguments passed to \code{\link{score}}}
}
\value{
ggplot2 heatmap visualization
}
\description{
Visualize the PPS matrix
}
\examples{
visualize_matrix(iris)
}
