% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_pps}
\alias{visualize_pps}
\title{Visualize the Predictive Power scores of the entire dataframe, or given a target}
\usage{
visualize_pps(
  df,
  y = NULL,
  color_value_high = "#08306B",
  color_value_low = "#FFFFFF",
  color_text = "#000000",
  ...
)
}
\arguments{
\item{df}{data.frame containing columns for x and y}

\item{y}{string, column name of target variable,
can be left \code{NULL} to visualize all X-Y PPS}

\item{color_value_high}{color used for upper limit of PPS gradient (high PPS)}

\item{color_value_low}{color used for lower limit of PPS gradient (low PPS)}

\item{color_text}{color used for text, best to pick high contrast with \code{color_value_high}}

\item{...}{any arguments passed to \code{\link{score}}}
}
\value{
ggplot2 vertical barplot or heatmap visualization
}
\description{
If \code{y} is specified, \code{visualize_pps} returns a barplot of the PPS of
every predictor on the specified target variable.
If \code{y} is not specified, \code{visualize_pps} returns a heatmap visualization
of the PPS for all X-Y combinations in a dataframe.
}
\examples{
visualize_pps(iris, y = 'Species')

visualize_pps(iris)

\dontrun{visualize_pps(mtcars, do_parallel = TRUE)}
}
