% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_predictors}
\alias{visualize_predictors}
\title{Visualize the PPS of all predictors of a target}
\usage{
visualize_predictors(df, y, color = "#08306B")
}
\arguments{
\item{df}{data.frame}

\item{y}{string, column name of target variable}

\item{color}{color used for highlighting high PPS}
}
\value{
ggplot2 vertical barplot visualization
}
\description{
Visualize the PPS of all predictors of a target
}
\examples{
visualize_predictors(mtcars, 'mpg')

visualize_predictors(iris, 'Species')
}
